requirements.txt
langchain-astradb == 0.5.2
langchain == 0.3.14
langchain-openai == 0.2.14
pypdf == 5.1.0
python-dotenv == 1.0.1
Flask == 3.1.0
datasets == 3.2.0